feat: v1-balances
