+++ server/routes/v1/__init__.py
@@ -1,9 +1,10 @@
 from apiflask import APIBlueprint
 
-from . import accounts, blocks, transactions
+from . import accounts, balances, blocks, transactions
 
 v1_bp = APIBlueprint("v1", __name__, url_prefix="/v1")
 
 v1_bp.register_blueprint(accounts.accounts_bp)
+v1_bp.register_blueprint(balances.balances_bp)
 v1_bp.register_blueprint(blocks.blocks_bp)
 v1_bp.register_blueprint(transactions.transactions_bp)
